Fix QwixxLogic unit test mock errors due to enforced types in Class
